Model-Serving-API-Project/model-API/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi[all]==0.116.1
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
joblib==1.5.1
scikit-learn==1.7.0
numpy==2.3.1
pandas==2.3.1
celery[redis]==5.3.6
redis==5.0.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
auth0-python==4.0.0
requests==2.32.3
xgboost>=1.7.5
ta>=0.10.2
pytest==8.4.1
pytest-asyncio==1.1.0
httpx==0.28.1